add service, refactor files
