+++ crudapi/src/main/java/com/jms/crudapi/book/Book.java
@@ -1,4 +1,4 @@
-package com.jms.crudapi.model;
+package com.jms.crudapi.book;
 
 import lombok.*;
