+++ crudapi/src/main/java/com/jms/crudapi/book/BookRepository.java
@@ -1,0 +1,7 @@
+package com.jms.crudapi.book;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookRepository extends JpaRepository<Book, Long>{
+
+}
