+++ crudapi/src/main/java/com/jms/crudapi/book/BookService.java
@@ -1,0 +1,5 @@
+package com.jms.crudapi.book;
+
+public class BookService {
+    
+}
