+++ crudapi/src/main/java/com/jms/crudapi/book/exception/ControllerExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.jms.crudapi.exception;
+package com.jms.crudapi.book.exception;
 
 import java.util.Date;
