+++ crudapi/src/test/java/com/jms/crudapi/BookUseCase.java
@@ -1,0 +1,5 @@
+package com.jms.crudapi;
+
+public class BookUseCase {
+
+}
